Add unit test for Other variant (#31)
* Add unit test for Other variant

* Fix deserialization

* Add ID link

* i32 -> i64

* bump version

+++ src/models/admin.rs
@@ -20,14 +20,14 @@ pub struct Admin {
 		with = "::serde_with::rust::double_option",
 		skip_serializing_if = "Option::is_none"
 	)]
-	pub attempts: Option<Option<i32>>,
+	pub attempts: Option<Option<i64>>,
 	#[serde(
 		rename = "expire_at",
 		default,
 		with = "::serde_with::rust::double_option",
 		skip_serializing_if = "Option::is_none"
 	)]
-	pub expire_at: Option<Option<i32>>,
+	pub expire_at: Option<Option<i64>>,
 }
 
 impl Admin {
